+++ source/tests/system/nisync_driver_api_tests.cpp
@@ -193,6 +193,62 @@ class NiSyncDriverApiTest : public ::testing::Test {
     return grpcStatus;
   }
 
+  ::grpc::Status call_SetAttributeViInt32(ViConstString terminalName, ViAttr attribute, ViInt32 value, ViStatus* viStatusOut)
+  {
+    ::grpc::ClientContext clientContext;
+    nisync::SetAttributeViInt32Request request;
+    nisync::SetAttributeViInt32Response response;
+    request.set_terminal_name(terminalName);
+    request.set_attribute(static_cast<nisync::NiSyncAttributes>(attribute));
+    request.set_value(value);
+    request.mutable_vi()->set_id(driver_session_->id());
+    auto grpcStatus = GetStub()->SetAttributeViInt32(&clientContext, request, &response);
+    *viStatusOut = response.status();
+    return grpcStatus;
+  }
+
+  ::grpc::Status call_GetAttributeViInt32(ViConstString terminalName, ViAttr attribute, ViInt32* valueOut, ViStatus* viStatusOut)
+  {
+    ::grpc::ClientContext clientContext;
+    nisync::GetAttributeViInt32Request request;
+    nisync::GetAttributeViInt32Response response;
+    request.set_terminal_name(terminalName);
+    request.set_attribute(static_cast<nisync::NiSyncAttributes>(attribute));
+    request.mutable_vi()->set_id(driver_session_->id());
+    auto grpcStatus = GetStub()->GetAttributeViInt32(&clientContext, request, &response);
+    *valueOut = response.value();
+    *viStatusOut = response.status();
+    return grpcStatus;
+  }
+
+  ::grpc::Status call_SetAttributeViString(ViConstString terminalName, ViAttr attribute, ViConstString value, ViStatus* viStatusOut)
+  {
+    ::grpc::ClientContext clientContext;
+    nisync::SetAttributeViStringRequest request;
+    nisync::SetAttributeViStringResponse response;
+    request.set_terminal_name(terminalName);
+    request.set_attribute(static_cast<nisync::NiSyncAttributes>(attribute));
+    request.set_value(value);
+    request.mutable_vi()->set_id(driver_session_->id());
+    auto grpcStatus = GetStub()->SetAttributeViString(&clientContext, request, &response);
+    *viStatusOut = response.status();
+    return grpcStatus;
+  }
+
+  ::grpc::Status call_GetAttributeViString(ViConstString terminalName, ViAttr attribute, std::string* valueOut, ViStatus* viStatusOut)
+  {
+    ::grpc::ClientContext clientContext;
+    nisync::GetAttributeViStringRequest request;
+    nisync::GetAttributeViStringResponse response;
+    request.set_terminal_name(terminalName);
+    request.set_attribute(static_cast<nisync::NiSyncAttributes>(attribute));
+    request.mutable_vi()->set_id(driver_session_->id());
+    auto grpcStatus = GetStub()->GetAttributeViString(&clientContext, request, &response);
+    *valueOut = response.value();
+    *viStatusOut = response.status();
+    return grpcStatus;
+  }
+
 private:
   std::shared_ptr<::grpc::Channel> channel_;
   std::unique_ptr<::nidevice_grpc::Session> driver_session_;