+++ source/codegen/metadata/nisync/CHANGES.md
@@ -8,3 +8,15 @@ with the session_repository.
 
 The following function parameters were changes to camel case so the generated name would be snake case.
 - `idQuery` parameter in `init`
+
+The following function parameters were changed to generate the correct parameter type.
+- `value` paramter in `GetAttributeViString`
+```
+            "direction": "out",
+            "name": "value",
+            'size': {
+                'mechanism': 'ivi-dance',
+                'value': 'bufferSize'
+            },
+            'type': 'ViChar[]'
+```